Separate concerns.
